Add files via upload
This file shows the thought process of this program for reading from a file of questions and answering the correct question. If the question was answered correctly, the program will print correct and count the number of correct answers at the end of program. This will ultimately print to the screen.